gnugrep: new pkg following upstream
It's a GNU package, so it goes to gnu/. Provide a choice between default
grep and GNU grep since admittedly, GNU grep is way faster. Also it's a
userspace program so the impact is not really the same as having
libstdc++ (and it allows for headache-less switching using kiss a).